README: Don't try to provide instructions for getting development-only dependencies.
Those instructions are not very reliable and hard to provide. go get
command is meant for getting dependencies and building a package, it's
not a good command for only getting dependencies (since it expects the
target to be a valid Go package).

People who are interested in doing development should be able to figure
out how to get dependencies that are missing if they run into it, so
it's enough to mention that extra dependencies may be required for
development.