Add Options.BodyTop field to provide top of page components.
For now, this will be used for providing things like global navigation
bar on top of pages, using the dynamic component HTML rendering with
context available.

This is the current general direction and in the future, this app will
be rewritten more and more to be component-based rather than
template-based.

Regenerate.