Remove Options.BaseState.
Options.BaseState was an experimental approach to providing multiple
arguments to the notificationsapp handler. It was an attempt at replacing
the multiple individual arguments such as RepoSpec, BaseURI, Context,
and potentially others. It was also there in case additional things
needed to be customized.

Now that I've learned more about the use cases, it's no longer needed,
and can be removed without loss of functionality. Doing so makes the
API of notificationsapp less difficult to use.

This is a breaking API change. Providing BaseState is no longer needed.