Remove Options.BaseURI; rely on context key parameter.
BaseURI is now expected to be provided via request context
value, using BaseURIContextKey.

Using context key to provide required parameter to an http.Handler is
a little unusual and experimental. There may be room for future
improvement. However, this is an improvement over the previous approach
of having custom funcs that must extract this information from the
request (which was typically implemented using same context key
anyway).

This hopefully makes the API simpler to use.

This is a breaking API change. BaseURI now must be provided as context
value on each request to notificationsapp handler.