Remove the need for users service.
The notifications service already performs authorization check (it has
to). Now that we're receiving errors from notification service directly
(instead of it going through ExecuteTemplate), it's possible to detect
and propagate permission errors from there. The users service was used
only for this purpose previously, and is no longer needed.

Returning permission errors is important so the caller can redirect to
/login when necessary.

Resolves the following TODO:

	// THINK: Try to let service take care of authorization check.