Hide "mark all read" button when all individual notifications get read.
Do this by doing some simple logic on the frontend to calculate how
many notifications are read out of how many there are in total for a
given RepoNotifications component.

Previously, I thought this would require re-rendering the entire
component. That might still be a good general change/direction to go in
the future, but this workaround is sufficient for the needs of
notificationsapp for now.

Regenerate.

Fixes https://dmitri.shuralyov.com/issues/github.com/shurcooL/notificationsapp/10.