Move assets/_data/script package to frontend.
It's a high quality Go package, one that just happens to be for
frontend use. Move it to a shorter, friendlier import path that doesn't
get ignored by ./... import path pattern (previously, it was ignored
because it was in a _data directory).

Move assets/_data to _data. This reduces nesting and makes it easier to
work on _data contents. assets package just happens to use some _data
files.

Move importPathToDir helper to the bottom for better readability and
consistency.